+++ services/workflows-service/src/workflow/workflow.service.ts
@@ -257,7 +257,10 @@ export class WorkflowService {
         ...workflow.context,
         documents: workflow.context?.documents?.map(
           (document: DefaultContextSchema['documents'][number]) => {
-            return addPropertiesSchemaToDocument(document);
+            return addPropertiesSchemaToDocument(
+              document,
+              workflow.workflowDefinition.documentsSchema,
+            );
           },
         ),
       },
@@ -804,8 +807,8 @@ export class WorkflowService {
       id: documentId,
     };
 
-    const newDocument = addPropertiesSchemaToDocument(document);
-    const propertiesSchema = newDocument?.propertiesSchema ?? {};
+    const documentSchema = addPropertiesSchemaToDocument(document, workflowDef.documentsSchema);
+    const propertiesSchema = documentSchema?.propertiesSchema ?? {};
     if (Object.keys(propertiesSchema)?.length) {
       let propertiesSchemaForValidation = propertiesSchema;
       if (!checkRequiredFields) {
@@ -814,7 +817,7 @@ export class WorkflowService {
       }
       const validatePropertiesSchema = ajv.compile(propertiesSchemaForValidation);
 
-      const isValidPropertiesSchema = validatePropertiesSchema(newDocument?.properties);
+      const isValidPropertiesSchema = validatePropertiesSchema(documentSchema?.properties);
 
       if (!isValidPropertiesSchema) {
         throw new BadRequestException(
@@ -830,7 +833,7 @@ export class WorkflowService {
     const updatedWorkflow = await this.updateContextById(
       workflowId,
       {
-        documents: [newDocument],
+        documents: [documentSchema],
       },
       [projectId],
     );