+++ sdks/workflow-node-sdk/CHANGELOG.md
@@ -4,7 +4,10 @@
 
 ### Patch Changes
 
-- @ballerine/workflow-core@0.5.23
+- d5cca1e8: added defintion validation
+- Updated dependencies [d5cca1e8]
+- Updated dependencies [f4b5c0df]
+  - @ballerine/workflow-core@0.5.23
 
 ## 0.5.22